+++ SVGVideoMaker/examples/EllipseArc.py
@@ -5,7 +5,7 @@ Test EllipseArc of SVGVideoMaker
 # region Imports
 from SVGVideoMaker.geo.animation import AnimationType
 from SVGVideoMaker.geo.arc import EllipseArc
-from SVGVideoMaker.geo import Point2D, SVG
+from SVGVideoMaker.geo import Point2D, SVG, Ellipse
 from SVGVideoMaker.video import Video
 # endregion Imports
 
@@ -18,7 +18,7 @@ width, height = 110, 180
 center = Point2D(width // 2, height // 2)
 full_turn = 360
 
-svg = SVG()
+svg = SVG(width=width, height=height)
 svg.set_view_box(Point2D(0, 0), Point2D(width, height))
 
 def create_column(colum, sens):
@@ -31,7 +31,7 @@ def create_column(colum, sens):
 		arc = EllipseArc(center=Point2D(colum, i * position_offset + int(position_offset / 2)),
 		                 radius=Point2D(8, 8),
 		                 start_angle=start_angle, end_angle=end_angle)
-		arc.animations.add_animation(seconds * fps, sens * full_turn, AnimationType.ANGLES)
+		arc.animations.add_animation(seconds * fps, AnimationType.ANGLES, value=sens * full_turn)
 
 		stroke_incrementer *= 1.5 if i <= 4 else 0.5
 		arc.set_style(stroke_color="black", stroke_width=1 + stroke_incrementer, stroke_linecaps="round")
@@ -46,34 +46,38 @@ def main():
 
 	# Fill center ellipse
 	ellipse_top = EllipseArc(Point2D(width // 2, position_offset // 2), Point2D(20, 5), 0, 180)
-	ellipse_top.animations.add_animation(seconds * fps, full_turn, AnimationType.ANGLES)
+	ellipse_top.animations.add_animation(seconds * fps, AnimationType.ANGLES, value=full_turn)
 
 	ellipse_top_mid = EllipseArc(Point2D(width // 2, height // 2 - height // 4), Point2D(5, 10), 135, 405)
-	ellipse_top_mid.animations.add_animation(seconds * fps, full_turn, AnimationType.ANGLES)
+	ellipse_top_mid.animations.add_animation(seconds * fps, AnimationType.ANGLES, value=full_turn)
+	ellipse_top_mid.add_translation(seconds * 2 * fps, 0, 0)
+	ellipse_top_mid.add_translation(round(seconds * 2.5 * fps), 0, height // 4)
 
 	ellipse_bottom = EllipseArc(Point2D(width // 2, height - position_offset // 2), Point2D(20, 5), 180, 360)
-	ellipse_bottom.animations.add_animation(seconds * fps, full_turn, AnimationType.ANGLES)
+	ellipse_bottom.animations.add_animation(seconds * fps, AnimationType.ANGLES, value=full_turn)
 
 	ellipse_bottom_mid = EllipseArc(Point2D(width // 2, height // 2 + height // 4), Point2D(5, 10), 315, 585)
-	ellipse_bottom_mid.animations.add_animation(seconds * fps, full_turn, AnimationType.ANGLES)
+	ellipse_bottom_mid.animations.add_animation(seconds * fps, AnimationType.ANGLES, value=full_turn)
+	ellipse_bottom_mid.add_translation(seconds * 2 * fps, 0, 0)
+	ellipse_bottom_mid.add_translation(round(seconds * 2.5 * fps), 0, -(height // 4))
 
 	ellipse_center = EllipseArc(center, Point2D(30, 20), 180, 0)
-	ellipse_center.animations.add_animation(seconds * fps, full_turn, AnimationType.ANGLES)
+	ellipse_center.animations.add_animation(seconds * fps, AnimationType.ANGLES, value=full_turn)
 
-	ellipse_center_center = EllipseArc(center, Point2D(10, 20), 0, 180)
-	ellipse_center_center.animations.add_animation(seconds * fps, full_turn, AnimationType.ANGLES)
+	ellipse_center_center = Ellipse(center, 10, 20)
+	ellipse_center_center.add_rotate(seconds * fps, value=360)
 
 	svg.append(ellipse_top, ellipse_top_mid, ellipse_bottom, ellipse_bottom_mid,
 	           ellipse_center, ellipse_center_center)
 
 	for element in svg:
-		element.animations.add_animation((seconds * 1) * fps, 1, AnimationType.OPACITY)
-		element.animations.add_animation((seconds * 2) * fps, 0, AnimationType.OPACITY)
-		element.animations.add_animation((seconds * 3) * fps, 1, AnimationType.OPACITY)
+		element.add_opacity((seconds * 1) * fps, 1)
+		element.add_opacity((seconds * 2) * fps, 0)
+		element.add_opacity((seconds * 3) * fps, 1)
 
-	video = Video(svg, fps=fps)
+	video = Video(svg, fps=fps, width=width*5, height=height*5)
 	print(svg.display_animations())
-	video.save_movie(name="ellipse_arc", max_time=seconds * 3 + 1)
+	video.save_movie(name="ellipse_arc", end=seconds * 3 + 1)
 
 
 if __name__ == '__main__':