+++ SVGVideoMaker/examples/Misc.py
@@ -4,14 +4,14 @@ Test different function of SVGVideoMaker
 
 # region Imports
 from SVGVideoMaker.video import Video
-from geo.animation import AnimationType
-from geo.circle import Circle
-from geo.svg import SVG
-from geo.group import Group
-from geo.polygon import Polygon
-from geo.point import Point
-from geo.point import Point2D
-from geo import Arc
+from SVGVideoMaker.geo.animation import AnimationType
+from SVGVideoMaker.geo.ellipse import Circle
+from SVGVideoMaker.geo.svg import SVG
+from SVGVideoMaker.geo.group import Group
+from SVGVideoMaker.geo.polygon import Polygon
+from SVGVideoMaker.geo.point import Point
+from SVGVideoMaker.geo.point import Point2D
+from SVGVideoMaker.geo import Arc
 # endregion Imports
 
 
@@ -21,6 +21,7 @@ def main():
     fps = 5
 
     svg = SVG()
+    svg.set_background_color("black")
 
     g = Group()
     arc = Arc(Point2D(50, 50), Point2D(0, 0), end=Point2D(250, 50))
@@ -30,115 +31,47 @@ def main():
 
     p = Point([-150, -150])
     square = Polygon.square(0, 10, 50)
-    svg.append(g, p, Polygon([Point([0, 0]), Point([0, 50]), Point([50, 50]), Point([50, 0])]), square)
     poly = Polygon([Point([0, 0]), Point([0, 50]), Point([50, 50]), Point([50, 0])])
-
-    pautre = Polygon([Point([35.1,16.51]),
-                     Point([35.1,16.51]),
-                     Point([45.1,16.51]),
-                     Point([54.9,16.51]),
-                     Point([54.9,17.49]),
-                     Point([44.9,17.49]),
-                     Point([35.1,17.49]),
-                     Point([35.1,17.49])
-                     ])
-
-    # svg.append(poly)
-    # svg.append(pautre)
+    svg.append(g, p, Polygon([Point([0, 0]), Point([0, 50]), Point([50, 50]), Point([50, 0])]), square, poly)
 
     # Make animation
     video = Video(svg, fps=fps)
 
     # Add all animations
-    p.animations.add_animation(10 * fps, Point([100, 50]))
-    p.animations.add_animation(20 * fps, Point([100, 50]))
-    p.animations.add_animation(10 * fps, 30, AnimationType.INFLATION)
-    pautre.animations.add_animation(50 * fps, 0, AnimationType.OPACITY)
+    p.add_translation(10 * fps, x=100, y=50)
+    p.add_translation(20 * fps, x=100, y=50)
+    p.animations.add_animation(10 * fps, AnimationType.INFLATION, value=30)
 
-    g.add_animation(20 * fps, Point([100, 50]))
+    g.add_animation(20 * fps, value=Point([100, 50]))
 
-    square.animations.add_animation(2 * fps, Point([100, 100]))
-    square.animations.add_animation(4 * fps, Point([-100, -100]))
-    square.animations.add_animation(10 * fps, Point([100, 100]))
-    square.animations.add_animation(11 * fps, Point([-100, -100]))
-    square.animations.add_animation(20 * fps, Point([100, 100]))
-
-    poly.animations.add_animation(1 * fps,
-                                  [Point([-10, -10]), Point([-10, 10]), Point([10, 10]), Point([10, -10])],
-                                  AnimationType.MODIFICATION
-                                  )
-
-    poly.animations.add_animation(5 * fps,
-                                  [Point([1, 1]), Point([1, 1]), Point([1, 1]), Point([1, 1])],
-                                  AnimationType.MODIFICATION
-                                  )
+    square.add_translation(2 * fps, x=100, y=100)
+    square.add_translation(4 * fps, 100, -100)
+    square.add_translation(10 * fps, 100, 100)
+    square.add_translation(11 * fps, -100, -100)
+    square.add_translation(20 * fps, 100, 100)
 
     revert_triangle = [Point([0, 10]), Point([-10, -10]), Point([10, -10])]
-    poly.animations.add_animation(10 * fps,
-                                  revert_triangle,
-                                  AnimationType.MODIFICATION
-                                  )
-
     nice_shape = [Point([10, 20]), Point([-10, 20]), Point([-20, 10]), Point([-20, -10]), Point([-10, -20]), Point([10, -20]), Point([20, -10]), Point([20, 10])]
-    poly.animations.add_animation(20 * fps,
-                                  nice_shape,
-                                  AnimationType.MODIFICATION
-                                  )
-    poly.animations.add_animation(22 * fps, revert_triangle, AnimationType.MODIFICATION)
-    poly.animations.add_animation(round(22.5 * fps), nice_shape, AnimationType.MODIFICATION)
-    poly.animations.add_animation(23 * fps, revert_triangle, AnimationType.MODIFICATION)
-    poly.animations.add_animation(round(23.25 * fps), nice_shape, AnimationType.MODIFICATION)
-    poly.animations.add_animation(round(23.5 * fps), revert_triangle, AnimationType.MODIFICATION)
-    poly.animations.add_animation(24 * fps, nice_shape, AnimationType.MODIFICATION)
-    poly.animations.add_animation(round(24.1 * fps), revert_triangle, AnimationType.MODIFICATION)
-    poly.animations.add_animation(round(24.2 * fps), nice_shape, AnimationType.MODIFICATION)
-    poly.animations.add_animation(30 * fps, revert_triangle, AnimationType.MODIFICATION)
 
-    poly.animations.add_animation(22 * fps, revert_triangle, AnimationType.MODIFICATION)
-    poly.animations.add_animation(30 * fps, nice_shape, AnimationType.MODIFICATION)
+    poly.add_modification(1 * fps, [Point([-10, -10]), Point([-10, 10]), Point([10, 10]), Point([10, -10])])
+    poly.add_modification(5 * fps, [Point([1, 1]), Point([1, 1]), Point([1, 1]), Point([1, 1])])
 
-    pautre.animations.add_animation(5 * fps, [
-        Point([25.1, 16.51]),
-        Point([35.1, 16.51]),
-        Point([45.1, 16.51]),
-        Point([54.9, 16.51]),
-        Point([54.9, 17.49]),
-        Point([44.9, 17.49]),
-        Point([34.9, 17.49]),
-        Point([25.1, 17.49])
-    ], AnimationType.MODIFICATION)
+    poly.add_modification(10 * fps, revert_triangle)
+    poly.add_modification(20 * fps, nice_shape)
+    poly.add_modification(22 * fps, revert_triangle)
+    poly.add_modification(round(22.5 * fps), nice_shape)
+    poly.add_modification(23 * fps, revert_triangle)
+    poly.add_modification(round(23.25 * fps), nice_shape)
+    poly.add_modification(round(23.5 * fps), revert_triangle)
+    poly.add_modification(24 * fps, nice_shape)
+    poly.add_modification(round(24.1 * fps), revert_triangle)
+    poly.add_modification(round(24.2 * fps), nice_shape)
+    poly.add_modification(30 * fps, revert_triangle)
 
-    pautre.animations.add_animation(10 * fps, [
-        Point([25.1, 16.51]),
-        Point([35.1, 16.51]),
-        Point([35.1, 16.51]),
-        Point([45.1, 16.51]),
-        Point([45.1, 16.51]),
-        Point([54.9, 16.51]),
-        Point([54.9, 17.49]),
-        Point([44.9, 17.49]),
-        Point([34.9, 17.49]),
-        Point([25.1, 17.49])
-    ], AnimationType.MODIFICATION)
+    poly.add_modification(22 * fps, revert_triangle)
+    poly.add_modification(30 * fps, nice_shape)
 
-    pautre.animations.add_animation(15 * fps, [
-        Point([25.1, 16.51]),
-        Point([35.1, 16.49]),
-        Point([35.1, 15.51]),
-        Point([44.9, 15.51]),
-        Point([45.1, 16.51]),
-        Point([54.9, 16.51]),
-        Point([54.9, 17.49]),
-        Point([44.9, 17.49]),
-        Point([34.9, 17.49]),
-        Point([25.1, 17.49])
-    ], AnimationType.MODIFICATION)
-
-    arc.animations.add_animation(50 * fps, Point2D(-250, -250))
-    arc.animations.add_animation(60 * fps, Point2D(50, 50))
-    print(poly.animations.display_animations())
-    print(arc.animations.display_animations())
-    video.save_movie(name="my_out", max_time=10)
+    video.save_movie(name="my_out")
 
 
 if __name__ == '__main__':