+++ SVGVideoMaker/geo/animation.py
@@ -13,15 +13,16 @@ class AnimationType(Enum):
     TRANSLATION = 1
     MODIFICATION = 2
     INFLATION = 3
-    OPACITY = 4
-    ANGLES = 5
+    ROTATION = 4
+    OPACITY = 5
+    ANGLES = 6
 
 class AnimationState(Enum):
     """
     Enumeration of different state of animation.
     """
     NEW = 0
-    SAME = 1
+    CURRENT = 1
     END = 2
 
 class Animation:
@@ -64,9 +65,21 @@ class Animation:
     def finish_animation(self):
         self.anim_computed = None # Clear element computed
 
-    def add_animation(self, frame, values, anim_type=AnimationType.TRANSLATION):
+    def add_animation(self, frame, anim_type=AnimationType.TRANSLATION, **kwargs):
+        """Add animation pn the good type and check if data is ok
+
+        Args:
+            frame     (int)          : The frame number.
+            anim_type (AnimationType): The type of animation to add.
+            **kwargs                 : The data to add. Attribute can be x, y or value.
+        """
         # Check the data type and do some checking operation before
-        # TODO
+        if len(kwargs) == 2:
+            from SVGVideoMaker.geo.point import Point2D
+            values = Point2D(kwargs["x"], kwargs["y"])
+        else:
+            values = kwargs["value"]
+
         # Add data if is good
         self.anims[anim_type][frame] = deepcopy(values) # To don't be affect about possible modification
         self.nb_frames = max(frame, self.nb_frames)
@@ -77,6 +90,7 @@ class Animation:
         """
         self.update_translation()
         self.update_inflation()
+        self.update_rotation()
         self.update_opacity()
         # Increment frame counter
         self.current_frame += 1
@@ -113,13 +127,15 @@ class Animation:
                 # No animation for this element
                 return AnimationState.END, (None, 0), (None, 0)
         else:  # We have already the next animation data
-            return AnimationState.SAME, (None, None), (None, None)
+            return AnimationState.CURRENT, (None, None), (None, None)
 
-    def update_generic(self, anim_type, apply):
-        state, (_, start_frame), (values, end_frame) = self.read_animation(anim_type) # Read values of Animation type
+    def update_generic(self, anim_type, apply, sub=False):
+        # Read values of Animation type
+        state, (start_val, start_frame), (end_val, end_frame) = self.read_animation(anim_type)
         if state is AnimationState.NEW:
             # Compute addition and memoize for optimization
-            self.current_values[anim_type] = (end_frame, values / (end_frame - start_frame))
+            value = end_val - start_val if sub else end_val
+            self.current_values[anim_type] = (end_frame, value / (end_frame - start_frame))
         if state is not AnimationState.END:
             apply(self.current_values[anim_type][Animation.VALUES])
 
@@ -129,14 +145,11 @@ class Animation:
     def update_inflation(self):
         self.update_generic(AnimationType.INFLATION, self.svg_el.apply_inflation)
 
+    def update_rotation(self):
+        self.update_generic(AnimationType.ROTATION, self.svg_el.apply_rotation)
+
     def update_opacity(self):
-        state, (old_opacity, start_frame), (new_opacity, end_frame) = self.read_animation(AnimationType.OPACITY)
-        if state is AnimationState.NEW:
-            # Compute and apply opacity
-            self.current_values[AnimationType.OPACITY] = (end_frame, (new_opacity - old_opacity) / (end_frame - start_frame))
-            self.svg_el.apply_opacity(self.current_values[AnimationType.OPACITY][Animation.VALUES])
-        elif state is AnimationState.SAME:
-            self.svg_el.apply_opacity(self.current_values[AnimationType.OPACITY][Animation.VALUES])
+        self.update_generic(AnimationType.OPACITY, self.svg_el.apply_opacity, sub=True)
 
     def reset(self):
         """
@@ -230,7 +243,7 @@ class ModificationAnimation(Animation):
             self.current_values[AnimationType.MODIFICATION] = (end_frame, computed)
             # Send the modification to add
             self.svg_el.apply_modification(computed)
-        elif state is AnimationState.SAME:
+        elif state is AnimationState.CURRENT:
             # Send modification
             self.svg_el.apply_modification(self.current_values[AnimationType.MODIFICATION][Animation.VALUES])